zab/pyproject.toml
[project]
name = "zab"
version = "0.1.0"
description = "ZAB (Zookeeper Atomic Broadcast) consensus implementation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "rich>=14.2.0",
]
